A dmenu patch to disable the keyboard input field.
The added -x option prevents any character insertion or paste mode,
removing the cursor at the same time.

This is helpful to make a choice with a few items passed through stdin,
so one is not seeing them erased by typing anything on the keyboard.

A simple example could be a confirmation prompt:
$> printf "No\nYes" | dmenu -x -p "Do you want to exit ?"